Timestamp should be double as the time is a long int internally.

+++ scopewindow.cpp
@@ -431,7 +431,7 @@ void ScopeWindow::stopRec() {
 
 void ScopeWindow::writeFile() {
 	if (!rec_file) return;
-	fprintf(rec_file, "%e", ((float)nsamples) / ((float)attysScan.getAttysComm(0)->getSamplingRateInHz()));
+	fprintf(rec_file, "%e", ((double)nsamples) / ((double)attysScan.getAttysComm(0)->getSamplingRateInHz()));
 	for (int n = 0; n < attysScan.getNAttysDevices(); n++) {
 		for (int i = 0; i < AttysComm::NCHANNELS; i++) {
 			float phy = unfiltDAQData[n][i];
